src/test/java/com/rips7/day/Day10Test.java
package com.rips7.day;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class Day10Test extends DayTest {

    private final Day10 day10 = new Day10();

    @Test
    @Override
    public void testPart1() {
        assertEquals("13140", day10.part1(readTestResourceLinesPt1()));
    }

    @Test
    @Override
    public void testPart2() {
        assertEquals(
            """
                ##..##..##..##..##..##..##..##..##..##..
                ###...###...###...###...###...###...###.
                ####....####....####....####....####....
                #####.....#####.....#####.....#####.....
                ######......######......######......####
                #######.......#######.......#######.....""",
            day10.part2(readTestResourceLinesPt2()));
    }

}
